Update ESLint config
- Using PropTypes.object and PropTypes.array make sense to me. Sometime
PropTypes.arrayOf and PropTypes.shape are too verbose
- arrow-body-style should be used with care. Omitting curly braces might
look good but in the long term it causes a lot of changes when used to
create functional components. A lot of times the component needs
additional code and we end up changing it to use curly braces. It better
to maintain functional component consistent from the beginning. Other
uses of inline arrow functions look better if we omit the curly braces